Refactor: Move indexedCount calculation to DownloadsView to optimize main thread
- Moved the expensive `indexedCount` calculation from `App.tsx` to `components/DownloadsView.tsx`.
- The calculation now only runs when the "Downloads" view is active, preventing blocking the main thread during navigation or syncing in other views.
- Replaced the spread operator with an explicit loop when pushing children to the stack to avoid potential "Maximum call stack size exceeded" errors with deep/large trees.
- Updated `MainView` and `DownloadsView` props accordingly.
